tests: declare max_slots int32_t
Internally we use int and int32_t. We should settle on int32_t.

+++ tests/ts_print_mt.c
@@ -69,7 +69,7 @@ int main(int argc, char **argv)
 	struct input_absinfo slot;
 #endif
 	int32_t user_slots = 0;
-	unsigned short max_slots = 1;
+	int32_t max_slots = 1;
 	int ret, i, j;
 	int read_samples = 1;
 	short non_blocking = 0;
